requirements.txt
# Core image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Emotion & smile detection (optional but recommended)
deepface>=0.0.79
tensorflow>=2.13.0        # DeepFace backend

# Eye detection via Face Mesh (optional but recommended)
mediapipe>=0.10.0

# Configuration
PyYAML>=6.0

# Reporting
# (html & csv use stdlib — no extra deps needed)

# Download helper for DNN face detection model
requests>=2.31.0
tqdm>=4.66.0
